fix: Fix highlight points in calibration table and plot
Due to some rounding and mapping errors, the calibration table and plot
were not keeping the highlighted points in sync correctly.
If you clicked the wrong point in the plot, the table
wouldn't be able to find the corresponding entry to highligh.

This probably can be improved more. Mutating these datastructures might cuase more
problems than its worth, perhaps we should just re-render
everything every time a change happens. At the least
this should get improved so there is some notion
of a "key" for every point, so we don't rely upon
matching points by value.