Don't open serial ports until scale is selected
The previous method of blindly trying to open each serial port as we
discover it, to test if it is valid, is pretty sloppy.
It froze the GUI as it attemped to open the port, and I think it could
be disruptive to other system resources that are trying to use those
ports. Now, the Scale class has a open() method that must
be called before any read()s, this is what typically actually opens
the Serial port. Now this Scale.open() method is called only when a
user actually selects a scale. If a user stupidly chooses a bad port,
then we'll get problems then.