Use "ScaleSearcher" interface
Before the global function `scale.availableScales()`
supplied the list of scales, but this wasn't very customizable.
Now you can plug in whatever list of scales you want by
implementing your own ScaleSearcher, and then registering that
class with Boa.

Before the "Random Generator" option in the list of scales
was hardcoded into the .ui file, and also hardcoded into the
logic of selecting scales in Boa. With this new interface the
PhonyScaleSearcher is just another way to generate a different
kind of Scale object.