Fix PhonyScale so it actually uses open() properly
Before, PhonyScale generated data all the way back to the time
when you started the program. On a real scale we only should
have data back to when we called open(). Plus, as it was if you
selected the PhonyScale everything would freeze from getting barfed
way too much data.