Remove PyBluez interface
The PyBluez API wasn't working that well, printing warnings
and taking a long time to search for devices, as well as being
a large maintenance burden.

It's not a big loss either: MacOS, Linux, and Windows all have
separate support for creating a serial device from a Bluetooth
connection. So just do that separately, and then the normal
`SerialScale` interface works just fine. Tested on both MacOS
and Raspberry Pi (Linux).