Overhaul how baudrate is set
The old `isinstance` check was fragile and relied upon the scale
inheriting from SerialScale. Instead, we should
just be checking for if the thing
can quack, instead of asking if it's a duck.
This makes it less fragile from SerialScale
getting renamed as well (this is how this error
was caught).

Also, let's make it so the baudrate is only applied to the current
scale, instead of all of them.