Improve how scales are selected
Now the avgValue LCD screen on the GUI actually shows "Oe" for error
if there is no scale currently selected.

This also improves the interface so that selecting a scale vs
searching for scale by name is separated. The searching by name should
actually happen in the UI, that will soon move there.